+++ R/req-options.R
@@ -56,7 +56,7 @@ req_user_agent <- function(req, string = NULL) {
     )
     string <- paste0(names(versions), "/", versions, collapse = " ")
   } else {
-    check_string(string, "`string`")
+    check_string(string)
   }
 
   req_options(req, useragent = string)
@@ -75,7 +75,7 @@ req_user_agent <- function(req, string = NULL) {
 #' request("http://example.com") %>% req_timeout(10)
 req_timeout <- function(req, seconds) {
   check_request(req)
-  check_number(seconds, "`seconds`")
+  check_number_decimal(seconds)
 
   if (seconds < 0.001) {
     abort("`timeout` must be >1 ms")
@@ -107,11 +107,7 @@ req_proxy <- function(req, url, port = NULL, username = NULL, password = NULL, a
     proxyuserpwd <- NULL
   }
 
-  if (!is.null(port)) {
-    if (!is_integerish(port)) {
-      abort("`port` must be a number")
-    }
-  }
+  check_number_whole(port, allow_null = TRUE)
 
   req_options(
     req,
@@ -169,11 +165,12 @@ req_verbose <- function(req,
                         redact_headers = TRUE) {
   check_request(req)
 
+  to_redact <- attr(req$headers, "redact")
   debug <- function(type, msg) {
     switch(type + 1,
       text =       if (info)        verbose_message("*  ", msg),
       headerOut =  if (header_resp) verbose_header("<- ", msg),
-      headerIn =   if (header_req)  verbose_header("-> ", msg, redact_headers),
+      headerIn =   if (header_req)  verbose_header("-> ", msg, redact_headers, to_redact = to_redact),
       dataOut =    if (body_resp)   verbose_message("<< ", msg),
       dataIn =     if (body_req)    verbose_message(">> ", msg)
     )
@@ -197,13 +194,13 @@ verbose_message <- function(prefix, x) {
   cli::cat_line(prefix, lines)
 }
 
-verbose_header <- function(prefix, x, redact = TRUE) {
+verbose_header <- function(prefix, x, redact = TRUE, to_redact = NULL) {
   x <- readBin(x, character())
   lines <- unlist(strsplit(x, "\r?\n", useBytes = TRUE))
 
   for (line in lines) {
     if (grepl(":", line, fixed = TRUE)) {
-      header <- headers_redact(as_headers(line), redact)
+      header <- headers_redact(as_headers(line), redact, to_redact = to_redact)
       cli::cat_line(prefix, cli::style_bold(names(header)), ": ", header)
     } else {
       cli::cat_line(prefix, line)