+++ R/req-policy.R
@@ -1,6 +1,6 @@
-req_policies <- function(.req, ...) {
-  check_request(.req)
-  .req$policies <- modify_list(.req$policies, ...)
+req_policies <- function(.req, ..., error_call = caller_env()) {
+  check_request(.req, call = error_call)
+  .req$policies <- modify_list(.req$policies, ..., error_call = error_call)
   .req
 }
 
@@ -20,14 +20,14 @@ req_policy_call <- function(req, name, args, default) {
   }
 }
 
-as_callback <- function(x, n, name) {
+as_callback <- function(x, n, name, error_call = caller_env()) {
   if (is.null(x)) {
     return(x)
   }
 
   x <- as_function(x)
   if (!inherits(x, "rlang_lambda_function") && length(formals(x)) != n) {
-    abort(glue("Callback {name}() must have {n} argument"))
+    abort(glue("Callback {name}() must have {n} argument"), call = error_call)
   }
   x
 }