+++ R/url.R
@@ -25,7 +25,7 @@
 #' url$query <- list(a = 1, b = 2, c = 3)
 #' url_build(url)
 url_parse <- function(url) {
-  check_string(url, "`url`")
+  check_string(url)
 
   # https://datatracker.ietf.org/doc/html/rfc3986#appendix-B
   pieces <- parse_match(url, "^(([^:/?#]+):)?(//([^/?#]*))?([^?#]*)(\\?([^#]*))?(#(.*))?")
@@ -68,9 +68,9 @@ url_parse <- function(url) {
   )
 }
 
-url_modify <- function(url, ...) {
+url_modify <- function(url, ..., error_call = caller_env()) {
   url <- url_parse(url)
-  url <- modify_list(url, ...)
+  url <- modify_list(url, ..., error_call = error_call)
   url_build(url)
 }
 
@@ -139,6 +139,10 @@ url_build <- function(url) {
     authority <- NULL
   }
 
+  if (!is.null(url$path) && !startsWith(url$path, "/")) {
+    url$path <- paste0("/", url$path)
+  }
+
   prefix <- function(prefix, x) if (!is.null(x)) paste0(prefix, x)
   paste0(
     url$scheme, if (!is.null(url$scheme)) ":",
@@ -162,9 +166,9 @@ query_parse <- function(x) {
   out
 }
 
-query_build <- function(x) {
+query_build <- function(x, error_call = caller_env()) {
   if (!is_list(x) || (!is_named(x) && length(x) > 0)) {
-    abort("Query must be a named list")
+    abort("Query must be a named list", call = error_call)
   }
 
   x <- compact(x)
@@ -177,7 +181,7 @@ query_build <- function(x) {
     abort(c(
       "Query parameters must be length 1 atomic vectors.",
       paste0("Problems: ", paste0(names(x)[bad_val], collapse =", "))
-    ))
+    ), call = error_call)
   }
 
   is_double <- map_lgl(x, is.double)